EOS-27288: CAS request should succeed even one put request is success (#1363)
Problem:
When the cluster is operating in a degraded mode, if we get a failure
for one of the DIX put requests( which goes to the failed node) the
whole index put operation returns a failure even though the other put
request are successful. This causes a failure in bucket creation and in
multipart upload operations.

Solution:
Return success from motr client for dix put request if we get success
from even a single request out of N requests. Report a failure only if
all the put requests fail.

Signed-off-by: Yatin Mahajan <yatin.mahajan@seagate.com>
Signed-off-by: Atul Deshmukh <atul.deshmukh@seagate.com>