scripts: enable build-prep-1node.sh run from src dir (#1370)
`build-prep-1node.sh` was designed to run as a standalone
script. Now, after we added it to motr, users may try to
run it from the source directory, but it does not support
it currently.

Solution: enhance build-prep-1node.sh so that it can
recognize when it is run from the source directory and
behave accordingly.

* install-build-deps: fix dep regression on CentOS 7
  python-devel dependency is missing since commit 4f5936af2.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>
Signed-off-by: Atul Deshmukh <atul.deshmukh@seagate.com>