scripts: fix issues with build prepare scripts (#1416)
With `--no-ansible` option it would still install
ansible. And when it installs it (even without the option)
it would not make sure epel repo is installed and updated.
This is a regression introduced at commit 86d2c997f and
here we fix it.
Also  Fixed not calling of build-prep-1node.sh in
build-prep-1node-cortx-mgw.sh

Reviewed-by: Madhavrao Vemuri <madhav.vemuri@seagate.com>
Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>
Signed-off-by: Atul Deshmukh <atul.deshmukh@seagate.com>