debian: fix install stage on arm64 (#1417)
The install stage of .deb package build was hardcoded to
x86_64 platform, it's fixed to work for any platform now.

* debian: fix docbook dependency on arm64

docbook-xml package is a dependency of docbook-xsl and
should be installed automatically, but somehow on arm64
it was not. Solution: add it explicitly to the deps list.

Note: docbook packages are used to generate man pages.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>
Signed-off-by: Atul Deshmukh <atul.deshmukh@seagate.com>